+++ client/Packages/com.beamable/Common/Runtime/Content/Federation.cs
@@ -17,5 +17,17 @@ namespace Beamable.Common.Content
 
 	[System.Serializable]
 	[Agnostic]
+	public class Namespace
+	{
+		[ContentField("name")]
+		public string Name;
+	}
+
+	[System.Serializable]
+	[Agnostic]
 	public class OptionalFederation : Optional<Federation> { }
+
+	[System.Serializable]
+	[Agnostic]
+	public class OptionalNamespace : Optional<Namespace>{}
 }